+++ crates/lexgen/src/lib.rs
@@ -44,7 +44,7 @@ pub fn lexer(input: TokenStream) -> TokenStream {
     let mut semantic_action_table = SemanticActionTable::new();
 
     let Lexer {
-        derives,
+        attrs,
         public,
         type_name,
         user_state_type,
@@ -157,7 +157,7 @@ pub fn lexer(input: TokenStream) -> TokenStream {
         type_name,
         token_type,
         public,
-        derives,
+        attrs,
     )
     .into()
 }